RFC: initial commit of Gdelta
This adds a new fugitive command, "Gdelta", which allows you to diff the
current working directory with any commit via:

:Gdelta <commit ID>

It does this by populating the quickfix list with the list of files
changed versus that commit, with the text of the quickfix lines being
the diff stat short output.

When run without an argument Gdelta will run against HEAD~1.

Some keyboard shortcuts from the Gstatus window have been replicated -
"D" opens a Gdiff with the older version of the file, "R" and "r"
refresh the quickfix list, etc.

I haven't added any documentation to fugitive.txt as of yet - I wanted
to get feedback on the functionality of Gstatus and on other people's
use cases before I did too much documentation work.